NF: Extract ModifierKeys to Binding class
Refactoring from 8078 for 6052
